refactor: :recycle: extract BaseCardEditor
Introduce packages/shared/src/ui-editor/base-editor.ts — an abstract
LitElement base for flow-card UI editors.  Provides hass/@state fields,
setConfig (with abstract configStruct), connectedCallback, page navigation,
render, _valueChanged, _computeLabelCallback, and styles.  Subclasses
supply configPages, generalSchema, advancedSchema, and _renderLegacyAlerts.

Migrate PowerFlowCardPlusEditor to extend BaseCardEditor.
